Use stream operator to report result code
The cast operator for KeyStoreNativeReturnCode and
KeyStoreNativeReturnCode will be removed later.
There already exists a stream operator to get
at the underlying value. Use that instead.

Bug: 119771891
Test: Built for walleye successfully.

+++ wifi/keystore/1.0/default/keystore.cpp
@@ -111,7 +111,7 @@ Return<void> Keystore::getPublicKey(const hidl_string& keyId, getPublicKey_cb _h
 
     auto export_result = future.get();
     if (!export_result.resultCode.isOk()) {
-        LOG(ERROR) << AT << "exportKey failed: " << int32_t(export_result.resultCode);
+        LOG(ERROR) << AT << "exportKey failed: " << export_result.resultCode;
         _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
         return Void();
     }
@@ -197,14 +197,14 @@ Return<void> Keystore::sign(const hidl_string& keyId, const hidl_vec<uint8_t>& d
 
     rc = KSReturn(error_code);
     if (!rc.isOk()) {
-        LOG(ERROR) << AT << "Keystore begin returned: " << int32_t(rc);
+        LOG(ERROR) << AT << "Keystore begin returned: " << rc;
         _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
         return Void();
     }
 
     OperationResult result = future.get();
     if (!result.resultCode.isOk()) {
-        LOG(ERROR) << AT << "begin failed: " << int32_t(result.resultCode);
+        LOG(ERROR) << AT << "begin failed: " << result.resultCode;
         _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
         return Void();
     }
@@ -224,7 +224,7 @@ Return<void> Keystore::sign(const hidl_string& keyId, const hidl_vec<uint8_t>& d
 
         rc = KSReturn(error_code);
         if (!rc.isOk()) {
-            LOG(ERROR) << AT << "Keystore update returned: " << int32_t(rc);
+            LOG(ERROR) << AT << "Keystore update returned: " << rc;
             _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
             return Void();
         }
@@ -232,7 +232,7 @@ Return<void> Keystore::sign(const hidl_string& keyId, const hidl_vec<uint8_t>& d
         result = future.get();
 
         if (!result.resultCode.isOk()) {
-            LOG(ERROR) << AT << "update failed: " << int32_t(result.resultCode);
+            LOG(ERROR) << AT << "update failed: " << result.resultCode;
             _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
             return Void();
         }
@@ -252,7 +252,7 @@ Return<void> Keystore::sign(const hidl_string& keyId, const hidl_vec<uint8_t>& d
             if (!KSReturn(error_code).isOk()) {
                 LOG(ERROR) << AT << "abort failed: " << error_code;
             } else if (!(rc = KSReturn(abortFuture.get().response_code())).isOk()) {
-                LOG(ERROR) << AT << "abort failed: " << int32_t(rc);
+                LOG(ERROR) << AT << "abort failed: " << rc;
             }
             _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
             return Void();
@@ -276,7 +276,7 @@ Return<void> Keystore::sign(const hidl_string& keyId, const hidl_vec<uint8_t>& d
 
     rc = KSReturn(error_code);
     if (!rc.isOk()) {
-        LOG(ERROR) << AT << "Keystore finish returned: " << int32_t(rc);
+        LOG(ERROR) << AT << "Keystore finish returned: " << rc;
         _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
         return Void();
     }
@@ -284,7 +284,7 @@ Return<void> Keystore::sign(const hidl_string& keyId, const hidl_vec<uint8_t>& d
     result = future.get();
 
     if (!result.resultCode.isOk()) {
-        LOG(ERROR) << AT << "finish failed: " << int32_t(result.resultCode);
+        LOG(ERROR) << AT << "finish failed: " << result.resultCode;
         _hidl_cb(KeystoreStatusCode::ERROR_UNKNOWN, {});
         return Void();
     }
